Initial commit
Create a utility that can get the metadata of a Blogger blog using the
Blogger v3 API.

The end goal is to be able to fetch existing blog posts from a given
blog, store the posts in a local database, and then replay them over
time into our own RSS feeds. This will allow us to gradually read
through a blog's archive.

The initial version will likely only support Blogger blogs, but anything
that has a reasonable API or is scrapeable might be supported in the
future.

To query the Blogger API, users need to create a Google Cloud API key[1]
and store it in a config file `~/.config/rss-replay/rss-replay.toml` (or
the equivalent path for non-Linux systems).

[1]: https://cloud.google.com/docs/authentication/api-keys#creating_an_api_key