Restrict visibility of blogger items
By taking a URL in blogger::get_posts(), we can remove the necessity for
outside callers to first call blogger::get_blog(). We can also restrict
the visibility of the blogger::Blog struct. Do so, so that we hide as
much implementation from callers as possible.

Once we're ready to add persistence, future commits will probably also
want to restrict blogger::get_posts(), and just provide a
blogger::scrape() function or similar.