Generate feeds incrementally
We don't want to write the full feed when we run the `generate`
subcommand. Instead we want to update the feed with the next oldest
entry. This way, we can grow the feed over time, gradually replaying the
entries out of the DB.

Implement this by simply reading the existing feed (if it exists) from
the filesystem, then popping the next oldest entry off the feed's DB
tree and appending it to the feed object. Then we write the updated feed
back to the filesystem.