scrape: remove expected target feed file
When we scrape a blog, remove any feed that might already be present.
This file is either not a feed, in which case we can't parse it, a feed
for a different blog, in which case we shouldn't combine it with the new
feed, or a feed for the same blog, in which case we will probably
generate duplicate entries.