Split feed generation into a separate function
Add a function for generating a single feed from the DB. This will make
it easier to generate a feed at scrape time, and may make it easier to
allow isolation between multiple feeds (so some feeds may successfully
update even if there are problems generating other feeds).

While we're at it, remove 'async' from the generation functions, since
these do not interact with reqwest and therefore do not need to be
async.