Add Blog trait
The main logic of scraping and saving posts to the DB doesn't need to
care about the type of blog involved. Start simplifying the
implementation by adding a Blog trait, which both the Blogger and
Wordpress modules can implement.

Move the generation of FeedData into the trait, and simplify the
existing get_feed() implementations. As we move more behavior into the
trait in later commits, we can hopefully provide a default
implementation for some of the common logic.