Move entry scraping to Blog trait
Continue refactoring Blogger and Wordpress implementations such that the
main logic doesn't need to care which implementation is used to get the
feed entries.

Replace the old get_feed() functions with a Blog method entries(). For
now, this is basically the same implementation as get_feed() with a few
small simplifications where items are able to be precomputed when the
Blog object is created.

While we're at it, hardcode the 1s delay when making multiple calls to
the Blogger API.

Future commits should further simplify the logic by precomputing more
items when creating the Blog objects.